\name{ec.Save}
\alias{ec.Save}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{TODO}
\description{
TODO}
\usage{
ec.Save(name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{name}{
TODO
}

}

\value{
Nothing
}
\author{jp.decorps@epiconcept.fr}


\examples{
#TODO
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
