\name{ec.summary}
\alias{ec.summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Descriptive Statistics : Summary of a variable}
\description{
Create a summary of the variable
}
\usage{
ec.summary(x, detail)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{TODO}
  \item{detail}{TODO}
}

\value{
Nothing
}
\author{jp.decorps@epiconcept.fr}


\examples{
data(mtcars)
ec.Use(mtcars)
ec.summary("cyl")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
