\name{ec.tabulate}
\alias{ec.tabulate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Descriptive Statistics : TODO}
\description{
Create a summary of the variable
}
\usage{
ec.tabulate(x, nolabel, missing)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{TODO}
  \item{nolabel}{TODO}
  \item{missing}{TODO}

}

\value{
Nothing
}
\author{jp.decorps@epiconcept.fr}


\examples{
#TODO
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
