\name{Tiramitsu}
\alias{Tiramitsu}
\docType{data}
\title{A TIAC Dataset}
\description{
TODO
}
\usage{data(Tiramitsu)}
\format{
  A data frame with 291 observations on the following 21 variables.
  \describe{
    \item{\code{ill}}{a numeric vector}
    \item{\code{dateonset}}{a Date}
    \item{\code{sex}}{a factor with levels \code{females} \code{males}}
    \item{\code{age}}{a numeric vector}
    \item{\code{tira}}{a numeric vector}
    \item{\code{tportion}}{a numeric vector}
    \item{\code{wmousse}}{a numeric vector}
    \item{\code{dmousse}}{a numeric vector}
    \item{\code{mousse}}{a numeric vector}
    \item{\code{mportion}}{a numeric vector}
    \item{\code{beer}}{a numeric vector}
    \item{\code{uniquekey}}{a numeric vector}
    \item{\code{redjelly}}{a numeric vector}
    \item{\code{fruitsalad}}{a numeric vector}
    \item{\code{tomato}}{a numeric vector}
    \item{\code{mince}}{a numeric vector}
    \item{\code{salmon}}{a numeric vector}
    \item{\code{horseradish}}{a numeric vector}
    \item{\code{chickenwin}}{a numeric vector}
    \item{\code{roastbeef}}{a numeric vector}
    \item{\code{pork}}{a numeric vector}
  }
}
\details{
TODO
}
\source{
TODO
}
\references{
TODO
}
\examples{
data(Tiramitsu)
## maybe str(Tiramitsu) ; plot(Tiramitsu) ...
}
\keyword{datasets}
